Simple prometheus API client
This is a fairly thin layer over Faraday to make using the Promethus API
easier and less error prone. It handles errors properly (unlike the
official promethus client (for example if a query contains an error nil
is returned; the response object and any error messages are all
discarded)) plus it uses dry-types and dry-monads to shore up the
safety even further.

At the moment the specs require a running server (specifically they
require rgh's server containing specific data) which is very far from
ideal and authentication probably won't be very pretty as it simply
passes everything through to Faraday.