Specify some default headers
This sets the User-Agent and Accept. The Accept header is completely
pointless becuase Prometheus always returns json and Loki returns json
when the query returns a 200 and plain text when the query returns 400.
Yah for consistence.

Either way it doesn't matter what the Accept header is Prometheus/Loki
just ignore it. However this library can at least do the right thing.