Remove the gemspec from the Gemfile
It turns out that having two gemspec entries in the Gemfile breaks the
gems built in the this repo so removing them seems like a good idea!

How I manage future development is a bridge that I will cross when I get
to it. Maybe someone will have fixed bundler. Yeah right.